# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TEST",
        "title": "API",
        "contact": {},
        "version": "1.7"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/countries/add": {
            "post": {
                "description": "This endpoint adds a new country.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Country"
                ],
                "summary": "Add a country",
                "parameters": [
                    {
                        "description": "Country details to add",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Country"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Country"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Message"
                        }
                    }
                }
            }
        },
        "/countries/delete": {
            "delete": {
                "description": "This endpoint deletes a country by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Country"
                ],
                "summary": "Delete a country",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country ID to delete",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Message"
                        }
                    }
                }
            }
        },
        "/countries/edit": {
            "put": {
                "description": "This endpoint edits an existing country.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Country"
                ],
                "summary": "Edit a country",
                "parameters": [
                    {
                        "description": "Country details to edit",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Country"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Country"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Message"
                        }
                    }
                }
            }
        },
        "/countries/get": {
            "get": {
                "description": "This endpoint retrieves a country by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Country"
                ],
                "summary": "Get a country",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country ID to retrieve",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Country"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Message"
                        }
                    }
                }
            }
        },
        "/countries/getall": {
            "get": {
                "description": "This endpoint retrieves all countries with pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Country"
                ],
                "summary": "List countries",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.ListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "olympy_api-gateway_genproto_country_service.Country": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "flag": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "olympy_api-gateway_genproto_country_service.ListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "countries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/olympy_api-gateway_genproto_country_service.Country"
                    }
                }
            }
        },
        "olympy_api-gateway_genproto_country_service.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}